remove unused and superceded headers, split out some functionality, and generally clean up the benchmarks
