+++ bench/syclblas/syclblas_benchmark_level1.cpp
@@ -23,7 +23,7 @@
  *
  **************************************************************************/
 
-#include "../blas_benchmark2.hpp"
+#include "../blas_benchmark.hpp"
 
 #include <interface/blas1_interface.hpp>
