+++ bench/syclblas/syclblas_benchmark_level2.cpp
@@ -23,7 +23,7 @@
  *
  **************************************************************************/
 
-#include "../blas_benchmark2.hpp"
+#include "../blas_benchmark.hpp"
 
 #include <interface/blas2_interface.hpp>
