+++ bench/syclblas/syclblas_benchmark_level3.cpp
@@ -23,8 +23,7 @@
  *
  **************************************************************************/
 
-// #include "../blas_benchmark.hpp"
-#include "../blas_benchmark2.hpp"
+#include "../blas_benchmark.hpp"
 
 #include <interface/blas1_interface.hpp>
 #include <interface/blas3_interface.hpp>
